+++ tests/testthat/test-warehousing-postgres.R
@@ -351,37 +351,52 @@ test_that("Ramses on PosgreSQL (system test)", {
     t = 0:5,
     patient_id = "99999999999",
     therapy_id = "5528fc41106bb48eb4d48bc412e13e67",
-    therapy_start = as.POSIXct("2015-08-07 10:27:00", tz = "Europe/London"),
-    therapy_end = as.POSIXct("2015-08-17 12:20:00", tz = "Europe/London"),
-    t_start = as.POSIXct(
-      c("2015-08-07 10:27:00", "2015-08-07 11:27:00", "2015-08-07 12:27:00",
-        "2015-08-07 13:27:00", "2015-08-07 14:27:00", "2015-08-07 15:27:00"), tz = "Europe/London"),
-    t_end = as.POSIXct(
-      c("2015-08-07 11:27:00", "2015-08-07 12:27:00", "2015-08-07 13:27:00", 
-        "2015-08-07 14:27:00", "2015-08-07 15:27:00", "2015-08-07 16:27:00"), tz = "Europe/London"),
+    therapy_start = structure(1438939620, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    therapy_end = structure(1439810400, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_start = structure(1438939620 + 0:5 * 3600, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_end = structure(c(1438939620 + 1:6 * 3600), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
     parenteral = 1L
   )
   test_expected_tail <- dplyr::tibble(
     t = 236:241,
     patient_id = "99999999999",
     therapy_id = "5528fc41106bb48eb4d48bc412e13e67",
-    therapy_start = as.POSIXct("2015-08-07 10:27:00", tz = "Europe/London"),
-    therapy_end = as.POSIXct("2015-08-17 12:20:00", tz = "Europe/London"),
-    t_start = as.POSIXct(
-      c("2015-08-17 06:27:00", "2015-08-17 07:27:00", "2015-08-17 08:27:00", 
-        "2015-08-17 09:27:00", "2015-08-17 10:27:00", "2015-08-17 11:27:00"), tz = "Europe/London"),
-    t_end = as.POSIXct(
-      c("2015-08-17 07:27:00", "2015-08-17 08:27:00", "2015-08-17 09:27:00", "2015-08-17 10:27:00", 
-        "2015-08-17 11:27:00", "2015-08-17 12:20:00"), tz = "Europe/London"),
+    therapy_start = structure(1438939620, tzone = "Europe/London", class = c("POSIXct", "POSIXt")),
+    therapy_end = structure(1439810400, tzone = "Europe/London", class = c("POSIXct", "POSIXt")),
+    t_start = structure(1438939620 + 236:241 * 3600, tzone = "Europe/London", class = c("POSIXct", "POSIXt")),
+    t_end = structure(c(1438939620 + 237:241 * 3600, 1439810400), tzone = "Europe/London", class = c("POSIXct", "POSIXt")),
     parenteral = 0L
   )
 
   expect_equal(head(test_output), test_expected_head)
   expect_equal(tail(test_output), test_expected_tail)
   expect_equal(
-    sum(difftime(test_output$t_end, test_output$t_start,units =  "hours")),
+    sum(difftime(test_output$t_end, test_output$t_start,units = "hours")),
     structure(241.883333333333, class = "difftime", units = "hours")
   )
+  
+  test_episode_extended <- TherapyEpisode(pq_conn, "5528fc41106bb48eb4d48bc412e13e67",
+                                          extend_table_start = 2)
+  test_output_extended <- longitudinal_table(test_episode_extended, collect = T)
+  
+  test_expected_head_extended <- dplyr::tibble(
+    t = -2:3,
+    patient_id = "99999999999",
+    therapy_id = "5528fc41106bb48eb4d48bc412e13e67",
+    therapy_start = structure(1438939620, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    therapy_end = structure(1439810400, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_start = structure(1438939620 + -2:3 * 3600, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_end = structure(c(1438939620 + -1:4 * 3600), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    parenteral = c(NA, 1L, 1L, 1L, 1L, 1L)
+  )
+  expect_equal(head(test_output_extended), test_expected_head_extended)
+  expect_equal(tail(test_output_extended), test_expected_tail)
+  expect_equal(
+    sum(difftime(test_output_extended$t_end, test_output_extended$t_start, units = "hours")),
+    structure(241.883333333333 + 2, class = "difftime", units = "hours") 
+  )
+  
+  # TherapyEpisode() method for MedicationRequest object
   
   test_medication_request <- MedicationRequest(pq_conn, "5528fc41106bb48eb4d48bc412e13e67")
   expect_is(test_medication_request, "MedicationRequest")
@@ -406,24 +421,48 @@ test_that("Ramses on PosgreSQL (system test)", {
     t = 117:122, 
     patient_id = "8258333156", 
     therapy_id = "f770855cf9d424c76fdfbc9786d508ac", 
-    therapy_start = structure(
-      c(1444239793, 1444239793, 1444239793, 1444239793, 
-        1444239793, 1444239793), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
-    therapy_end = structure(
-      c(1444681333, 1444681333, 1444681333, 1444681333, 
-        1444681333, 1444681333), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
-    t_start = structure(
-      c(1444660993, 1444664593, 1444668193, 1444671793, 
-        1444675393, 1444678993), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
-    t_end = structure(
-      c(1444664593, 1444668193, 1444671793, 1444675393, 
-        1444678993, 1444681333), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    therapy_start = structure(1444239793, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    therapy_end = structure(1444681333, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_start = structure(1444239793 + 117:122 * 3600, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_end = structure(c(1444239793 + 118:122 * 3600, 1444681333), tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
     parenteral = 0L
   )
   expect_equal(head(longitudinal_table(test_episode, collect = TRUE)), 
                test_expected_head)
   expect_equal(tail(longitudinal_table(test_episode, collect = TRUE)), 
                test_expected_tail_second_therapy_episode)
+  
+  test_episode_extended <- TherapyEpisode(
+    conn = pq_conn, 
+    id = c("f770855cf9d424c76fdfbc9786d508ac", 
+           "5528fc41106bb48eb4d48bc412e13e67"),
+    extend_table_start = 2
+  )
+  
+  test_expected_head_second_episode <- dplyr::tibble(
+    t = -2:3, 
+    patient_id = "8258333156", 
+    therapy_id = "f770855cf9d424c76fdfbc9786d508ac", 
+    therapy_start = structure(1444239793, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    therapy_end = structure(1444681333, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_start = structure(1444239793 + -2:3 * 3600, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    t_end = structure(1444239793 + -1:4 * 3600, tzone = "Europe/London", class = c("POSIXct", "POSIXt")), 
+    parenteral = c(NA, 1L, 1L, 1L, 1L, 1L)
+  )
+  
+  expect_equal(
+    head(longitudinal_table(test_episode_extended, collect = TRUE)),
+    test_expected_head_extended
+  )
+  expect_equal(
+    tail(longitudinal_table(test_episode_extended, collect = TRUE)), 
+    test_expected_tail_second_therapy_episode
+  )
+  expect_equal(
+    head(dplyr::filter(longitudinal_table(test_episode_extended, collect = TRUE),
+                       .data$therapy_id == "f770855cf9d424c76fdfbc9786d508ac")), 
+    test_expected_head_second_episode
+  )
   
   # TherapyEpisode .longitudinal_table_completeness_check -------------------------------------
   
@@ -849,7 +888,6 @@ test_that("Encounter class on Postgres", {
     ))
   # Encounter ------------------------------------------------------------------
   
-  # Single IVPO change pt 99999999999
   test_encounter <- Encounter(pq_conn, "3968305736")
   test_output <- longitudinal_table(test_encounter, collect = T)
   test_expected_head <- dplyr::tibble(
@@ -876,6 +914,26 @@ test_that("Encounter class on Postgres", {
   expect_equal(
     as.numeric(sum(difftime(test_output$t_end, test_output$t_start,units =  "days"))),
     sum(collect(test_encounter)[["ramses_bed_days"]])
+  )
+  
+  test_encounter_extended <- Encounter(pq_conn, "3968305736", extend_table_start = 2)
+  test_output_extended <- longitudinal_table(test_encounter_extended, collect = T)
+  
+  test_expected_head_extended <- dplyr::tibble(
+    t = -2:3,
+    patient_id = "99999999999",
+    encounter_id = "3968305736",
+    admission_date = structure(1486982520, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    discharge_date = structure(1487932800, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    t_start = structure(1486982520 + -2:3*3600, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    t_end = structure(1486982520 + -1:4*3600, tzone = "UTC", class = c("POSIXct", "POSIXt"))
+  )
+  
+  expect_equal(head(test_output_extended), test_expected_head_extended)
+  expect_equal(tail(test_output_extended), test_expected_tail)
+  expect_equal(
+    as.numeric(sum(difftime(test_output_extended$t_end, test_output_extended$t_start, units =  "days"))),
+    sum(collect(test_encounter)[["ramses_bed_days"]]) + 2/24
   )
   
   # 2+ Encounters --------------------------------------------------------------
@@ -939,14 +997,30 @@ test_that("Encounter class on Postgres", {
   ) %>% 
     longitudinal_table(collect = T)
   
+  last_temp_extended_1h <- clinical_feature_last(
+    Encounter(pq_conn, "9278078393", extend_table_start = 1),
+    observation_code = "8310-5",
+    hours = 24
+  ) %>% 
+    longitudinal_table(collect = T)
+  
   expect_equal(
     last_temp$last_temperature_24h[1:5],
     c(NA, NA, NA, 35.7, 35.7)
   )
   expect_equal(
+    last_temp_extended_1h$last_temperature_24h[1:6],
+    c(NA, NA, NA, NA, 35.7, 35.7)
+  )
+  expect_equal(
     last_temp$last_temperature_24h[21:25],
     c(37.1, 37.1, 37.1, 37.1, 37.1)
   )
+  expect_equal(
+    last_temp_extended_1h$last_temperature_24h[22:26],
+    c(37.1, 37.1, 37.1, 37.1, 37.1)
+  )
+  
   rm(last_temp)
   
   last_temp_2encounters <- clinical_feature_last(